Planets prop now passes down from app
