+++ client/src/containers/PlanetsContainer.js
@@ -1,19 +1,11 @@
 import { useState, useEffect } from 'react';
 import PlanetsList from "../components/Home/PlanetsList";
 
-const PlanetsContainer = () => {
-  const [planets, setPlanets] = useState([]);
+const PlanetsContainer = ({planets}) => {
+  
   const [selectedPlanets, setSelectedPlanets] = useState([]);
 
-  useEffect(() => {
-    getPlanets();
-  }, []);
-
-  const getPlanets = () => {
-    fetch("http://127.0.0.1:5000/api/planets")
-      .then(results => results.json())
-      .then(data => setPlanets(data));
-  };
+  
 
   const onPlanetClick = (planetName) => {
     const names = selectedPlanets.includes(planetName);
